Create Session-Validator.php
Code has been around for a long time; just finally decided to throw it on my github under examples;

Works fairly easily:
Make sure it's connected to a database where the username and password have been authenticated;
If the user is successfully validated they are moved to a new location;  Otherwise they are thrown away.

To be validated an individual 16 digit key is created and hashed alongside the username and password then an additional 2 digit number is thrown as well.

If the 16 digit key equals the 2 digit number then we hash the new number alongside the pre-existing username and password.  If they're both confirmed through this script then we allow them to pass, otherwise they cannot enter through this new point of security validation.